Crate pallet_evm_precompile_proxy

Structs

EvmSubCall

OnlyIsProxy

OnlyIsProxyAndProxy

ProxyPrecompile

A precompile to wrap the functionality from pallet-proxy.

Enums

ProxyPrecompileCall

Constants

CALL_DATA_LIMIT

Traits

EvmProxyCallFilter

A trait to filter if an evm subcall is allowed to be executed by a proxy account. This trait should be implemented by the ProxyType type configured in pallet proxy.