Trait pallet_evm_precompile_proxy::EvmProxyCallFilter

pub trait EvmProxyCallFilter: Sized + Send + Sync {
    fn is_evm_proxy_call_allowed(
        &self,
        _call: &EvmSubCall,
        _recipient_has_code: bool
    ) -> bool { ... }
}

A trait to filter if an evm subcall is allowed to be executed by a proxy account. This trait should be implemented by the ProxyType type configured in pallet proxy.

Provided Methods

fn is_evm_proxy_call_allowed(
&self,
_call: &EvmSubCall,
_recipient_has_code: bool
) -> bool

If returns false, then the subcall will not be executed and the evm transaction will revert with error message “CallFiltered”.